mainscreen done
add animation
title flicker
full rebuild
