+++ src/App.js
@@ -5,21 +5,19 @@ import { Navigation } from "modules/Navigation"
 import { Projects } from "modules/Projects"
 import { Technologies } from "modules/Technologies"
 import { Footer } from "./modules/Footer"
-import { MainScreen } from "./modules/MainScreen"
 
 function App() {
     return (
         <div>
-            {/* <Navigation /> */}
+            <Navigation />
             <div>
-                <MainScreen />
-                {/* <Main /> */}
-                {/* <About />
+                <Main />
+                <About />
                 <Projects />
                 <Technologies />
-                <Form /> */}
+                <Form />
             </div>
-            {/* <Footer /> */}
+            <Footer />
         </div>
     )
 }
