+++ src/assets/images/MainSection/index.js
@@ -1,0 +1,5 @@
+import planet from "./planet.webp"
+import Bsatellite from "./big-satellite.webp"
+import satellite from "./satellite.webp"
+
+export { Bsatellite, satellite, planet }
