--- src/modules/Main/components/BgAnimation/index.js
@@ -1,3 +1,0 @@
-import BgAnimation from "./BgAnimation"
-
-export { BgAnimation }
