+++ src/modules/Main/components/Hero/Hero.module.css
@@ -1,21 +1,91 @@
-.leftSection {
+.content {
     width: 100%;
+    padding-left: 40px;
 }
 
-@media (max-width:480px) {
-    .leftSection {
-        width: 80%;
-        display: flex;
-        flex-direction: column;
-        margin: 0 auto;
+.tagline {
+    font-weight: 500;
+    letter-spacing: 0.8px;
+    padding: 8px 10px;
+    background: linear-gradient(90.21deg, rgba(170, 54, 124, 0.502) -5.91%, rgba(74, 47, 189, 0.5) 111.58%);
+    border: none;
+    border-radius: 5px;
+    font-size: 20px;
+    margin-bottom: 16px;
+    display: inline-block;
+}
+
+.greeting {
+    max-width: fit-content;
+    font-size: clamp(44px, 4vw + 5px, 62px);
+    font-weight: 700;
+    letter-spacing: 0.8px;
+    line-height: 1;
+    margin-bottom: 20px;
+    display: block;
+    align-items: center;
+    justify-content: center;
+}
+
+.description {
+    max-width: 500px;
+    color: var(--gray-color);
+    font-size: 18px;
+    letter-spacing: 1px;
+    line-height: 1.5em;
+}
+
+.icon {
+    margin-left: 10px;
+}
+
+.connect {
+    font-weight: 700;
+    font-size: 20px;
+    margin: 30px 0;
+    letter-spacing: 0.8px;
+    display: flex;
+    align-items: center;
+    transition: 0.3s ease-in-out;
+    width: fit-content;
+    padding-left: 0;
+}
+
+.connect:hover {
+    margin-left: 25px;
+}
+
+@media (max-width: 960px) {
+    .content {
+        padding: 0 auto;
     }
 }
 
 @media (max-width: 650px) {
-    .leftSection {
-        width: 100%;
-        display: flex;
-        flex-direction: column;
-        margin: 0 auto;
+    .tagline {
+        display: none;
+    }
+
+    .greeting {
+        margin-top: 40px;
+        font-size: clamp(40px, 3vw + 10px, 44px);
+    }
+}
+
+@media (max-width: 480px) {
+    .tagline {
+        font-size: 16px;
+    }
+
+    .content {
+        padding: 20px;
+    }
+
+    .greeting {
+        font-size: clamp(32px, 5vw + 10px, 40px);
+    }
+
+    .description {
+        font-size: clamp(12px, 4vw + 1px, 18px)
     }
 }
