+++ src/modules/Main/components/PlanetAnimation/PlanetAnimation.module.css
@@ -1,0 +1,99 @@
+.imgContainer {
+    max-width: 500px;
+    max-height: fit-content;
+    animation: updown 10s linear infinite;
+    display: flex;
+    flex-shrink: 0;
+    justify-content: center;
+    align-items: center;
+}
+
+@keyframes updown {
+    0% {
+        transform: translateY(0);
+    }
+
+    50% {
+        transform: translateY(10px);
+    }
+
+    100% {
+        transform: translateY(0);
+    }
+}
+
+.planet {
+    width: 80%;
+    position: relative;
+    z-index: 1;
+    transform: rotate(155deg);
+    border-radius: 50%;
+    box-shadow: rgba(103, 44, 103, 0.756) 0px 7px 29px 0px;
+}
+
+.planetContainer {
+    width: 100%;
+}
+
+
+.Bsatellite {
+    position: absolute;
+    width: 10%;
+    border-radius: 50%;
+    animation: orbit 220s linear infinite;
+    -webkit-filter: drop-shadow(0px 0px 0px #000);
+    filter: drop-shadow(-2px 0px 2px #0000009e);
+    z-index: 1;
+    margin: auto;
+    top: 0;
+    left: 0;
+    right: 0;
+    bottom: 0;
+}
+
+@keyframes orbit {
+
+    0% {
+        transform: rotate(0deg) translateX(250px) translateY(0);
+    }
+
+    25% {
+        transform: rotate(90deg) translateX(0px) translateY(200px);
+    }
+
+    50% {
+        transform: rotate(180deg) translateX(-250px) translateY(0);
+    }
+
+    75% {
+        transform: rotate(270deg) translateX(0px) translateY(-300px);
+    }
+
+    100% {
+        transform: rotate(360deg) translateX(250px) translateY(0);
+    }
+}
+
+
+.satellite {
+    position: absolute;
+    width: 3%;
+    border-radius: 50%;
+    animation: satelliteOrb 35s linear infinite;
+    margin: auto;
+    z-index: 1;
+    top: 0;
+    left: 0;
+    right: 0;
+    bottom: 0;
+}
+
+@keyframes satelliteOrb {
+    from {
+        transform: rotate(360deg) translateX(220px);
+    }
+
+    to {
+        transform: rotate(0deg) translateX(220px);
+    }
+}
