+++ src/modules/Main/components/PlanetAnimation/index.js
@@ -1,0 +1,1 @@
+export { PlanetAnimation } from "./PlanetAnimation"
