--- src/modules/MainScreen/MainScreen.module.css
@@ -1,224 +1,0 @@
-.wrapper {
-    width: 100vw !important;
-    min-height: 100vh;
-    color: white;
-    background-image: url('../../assets/images/other/main-bg.png');
-    background-position: top center;
-    background-size: cover;
-    background-repeat: no-repeat;
-    display: flex;
-    justify-content: center;
-    align-items: center;
-    overflow: hidden;
-}
-
-.container {
-    width: 1200px;
-    min-height: 600px;
-    display: flex;
-    align-items: center;
-}
-
-.content {
-    width: 100%;
-    padding-left: 40px;
-}
-
-.tagline {
-    font-weight: 500;
-    letter-spacing: 0.8px;
-    padding: 8px 10px;
-    background: linear-gradient(90.21deg, rgba(170, 54, 124, 0.502) -5.91%, rgba(74, 47, 189, 0.5) 111.58%);
-    border: none;
-    border-radius: 5px;
-    font-size: 20px;
-    margin-bottom: 16px;
-    display: inline-block;
-}
-
-.greeting {
-    max-width: fit-content;
-    font-size: clamp(44px, 4vw + 5px, 62px);
-    font-weight: 700;
-    letter-spacing: 0.8px;
-    line-height: 1;
-    margin-bottom: 20px;
-    display: block;
-}
-
-.position {
-    color: rgba(170, 54, 124, 0.5);
-}
-
-.description {
-    max-width: 500px;
-    color: #B8B8B8;
-    font-size: 18px;
-    letter-spacing: 0.8px;
-    line-height: 1.5em;
-}
-
-.connect {
-    color: #fff;
-    font-weight: 700;
-    font-size: 20px;
-    margin-top: 30px;
-    letter-spacing: 0.8px;
-    display: flex;
-    align-items: center;
-    transition: 0.3s ease-in-out;
-    width: fit-content;
-}
-
-.connect:hover {
-    margin-left: 25px;
-}
-
-.imgContainer {
-    max-width: 500px;
-    max-height: fit-content;
-    animation: updown 10s linear infinite;
-    display: flex;
-    flex-shrink: 0;
-    justify-content: center;
-    align-items: center;
-
-}
-
-@keyframes updown {
-    0% {
-        transform: translateY(0);
-    }
-
-    50% {
-        transform: translateY(10px);
-    }
-
-    100% {
-        transform: translateY(0);
-    }
-}
-
-.planet {
-    width: 100%;
-    position: relative;
-    z-index: 1;
-    transform: rotate(155deg);
-    border-radius: 50%;
-}
-
-.planetContainer {
-    width: 100%;
-}
-
-
-.Bsatellite {
-    position: absolute;
-    width: 10%;
-    border-radius: 50%;
-    animation: orbit 220s linear infinite;
-    -webkit-filter: drop-shadow(0px 0px 0px #000);
-    filter: drop-shadow(-2px 0px 2px #0000009e);
-    z-index: 1;
-    margin: auto;
-    top: 0;
-    left: 0;
-    right: 0;
-    bottom: 0;
-}
-
-@keyframes orbit {
-
-    0% {
-        transform: rotate(0deg) translateX(250px) translateY(0);
-    }
-
-    25% {
-        transform: rotate(90deg) translateX(0px) translateY(200px);
-    }
-
-    50% {
-        transform: rotate(180deg) translateX(-250px) translateY(0);
-    }
-
-    75% {
-        transform: rotate(270deg) translateX(0px) translateY(-300px);
-    }
-
-    100% {
-        transform: rotate(360deg) translateX(250px) translateY(0);
-    }
-}
-
-
-.satellite {
-    position: absolute;
-    width: 3%;
-    border-radius: 50%;
-    animation: satelliteOrb 70s linear infinite;
-    margin: auto;
-    z-index: 1;
-    top: 0;
-    left: 0;
-    right: 0;
-    bottom: 0;
-}
-
-@keyframes satelliteOrb {
-    from {
-        transform: rotate(360deg) translateX(220px);
-    }
-
-    to {
-        transform: rotate(0deg) translateX(220px);
-    }
-}
-
-
-
-
-
-
-
-@media (max-width: 960px) {
-    .container {
-        flex-direction: column-reverse;
-    }
-
-    .content {
-        border: 1px solid green;
-        padding: 0 auto;
-    }
-}
-
-@media (max-width: 650px) {
-    .content {
-        display: flex;
-        flex-direction: column;
-        align-items: center;
-        padding-left: 0;
-    }
-
-    .greeting {
-        font-size: clamp(40px, 3vw + 10px, 44px);
-    }
-
-    .description {
-        text-align: center;
-        padding: 0 30px;
-    }
-}
-
-@media (max-width: 480px) {
-    .tagline {
-        font-size: 16px;
-    }
-
-    .content {
-        padding: 20px;
-    }
-
-    .greeting {
-        font-size: clamp(28px, 5vw + 10px, 40px);
-    }
-}
