--- src/modules/MainScreen/index.js
@@ -1,1 +1,0 @@
-export { MainScreen } from "./MainScreen"
