+++ src/modules/common/ui/TextFlicker/TextFlicker.js
@@ -1,0 +1,50 @@
+import React, { useState, useEffect } from "react"
+import styles from "./TextFlicker.module.css"
+
+export function TextFlicker({ toRotate }) {
+    const [loopNum, setLoopNum] = useState(0)
+    const [isDeleting, setIsDeleting] = useState(false)
+    const [text, setText] = useState("")
+    const [delta, setDelta] = useState(300 - Math.random() * 100)
+    const [index, setIndex] = useState(1)
+    const period = 2000
+
+    useEffect(() => {
+        let ticker = setInterval(() => {
+            tick()
+        }, delta)
+
+        return () => {
+            clearInterval(ticker)
+        }
+    }, [text])
+
+    const tick = () => {
+        let i = loopNum % toRotate.length
+        let fullText = toRotate[i]
+        let updatedText = isDeleting
+            ? fullText.substring(0, text.length - 1)
+            : fullText.substring(0, text.length + 1)
+
+        setText(updatedText)
+
+        if (isDeleting) {
+            setDelta(prevDelta => prevDelta / 2)
+        }
+
+        if (!isDeleting && updatedText === fullText) {
+            setIsDeleting(true)
+            setIndex(prevIndex => prevIndex - 1)
+            setDelta(period)
+        } else if (isDeleting && updatedText === "") {
+            setIsDeleting(false)
+            setLoopNum(loopNum + 1)
+            setIndex(1)
+            setDelta(500)
+        } else {
+            setIndex(prevIndex => prevIndex + 1)
+        }
+    }
+
+    return <span className={styles.position}>{text}</span>
+}
