+++ src/modules/common/ui/TextFlicker/index.js
@@ -1,0 +1,1 @@
+export { TextFlicker } from "./TextFlicker"
